Add files via upload
With the Adafruits "MAX31865" Module it is possible to measure temperatures with high accuracy (within ±0.1°C), from -40°C up to 750°C, here converted to a readable format with a DIY-Arduino-Board and a small OLED (128x64px). The MAX31865 contains a relative precise 15-Bit ADC, so its precision is greater than the peripherals (cable-losses, etc).
I use it to control the measures from NTC3950 measured f.ex. on 3D-Printers extruders, heaters (oven), etc.